requirements.txt
tensorflow
tensorflow-gpu
pillow
